requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
PyMySQL==1.1.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
tqdm==4.66.1
joblib==1.3.2

